requirements.txt
langgraph==0.6.6
langchain==0.3.27
langchain-openai==0.3.32
langchain-ibm==0.3.18
langchain-mcp-adapters==0.1.9
mcp==1.13.1

python-dotenv==1.1.1
streamlit==1.49.1
